% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_import.R
\name{bulk_import}
\alias{bulk_import}
\title{bulk import function}
\usage{
bulk_import(folder_path, merge = FALSE, file_type = "txt")
}
\arguments{
\item{folder_path}{= path of the file containing the raw data (i.e. 1, and 0's not "YES" and "NO"). The  file format currently accepted is .csv and .txt}

\item{merge:}{Can take on either two values. False: (the default) returns all imported datasets. TRUE: bulk_import attempts to return a single file containing all of the individual datasets. This currently only works if all datasets contain all of the same variables.}
}
\value{
Returns individual loaded dataset OR a single combined dataset and the raw labels.
}
\description{
A function to import multiple .txt or .csv files into R.
}
\examples{
bulk_import(folder_path = "C:/Users/jprendez/Desktop/bulk_import2", merge=TRUE, file_type="txt")
bulk_import()
}
\author{
Jordan L. Prendez, \email{jordanprendez@gmail.com}
}
\keyword{IR}
