% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{A simulated dataset containing 500 student records.}
\format{A data frame with 500 rows and 4 variables:
\describe{
  \item{IDType}{Unlabeled IDType}
  \item{Degree}{Unlabeled degree value}
  \item{Gender}{Unlabeled Gender value}
  \item{USCitizen}{Unlabeled USCitizen value}
  ...
}}
\usage{
example_dataset
}
\description{
A simulated dataset containing 500 student records.
}
\keyword{datasets}
