% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USM_labels.R
\name{usm_labels}
\alias{usm_labels}
\title{usm_labels function}
\usage{
usm_labels(dataset = NULL, label_values = TRUE, label_variables = TRUE,
  labels_output = FALSE)
}
\arguments{
\item{dataset}{= name of the object (file) that you wish to rename with USM labels.}

\item{label_values}{= TRUE (default). TRUE outputs your dataset with MHEC value labels attached.}

\item{label_variables}{= TRUE (default). TRUE outputs your dataset with MHEC variable labels attached.}

\item{labels_output}{= FALSE (default). TRUE outputs the labels used in the function.}
}
\value{
Returns dataset with USM value labels and or variable names attached (depending on selected option).
}
\description{
A function to attach USM-MHEC variable labels to a file currently loaded to memory.
Specifically, this file renames database style names to those more appropriate for reading and or tables.
}
\examples{
usm_labels(dataset=my_dataset, label_values = TRUE, label_variables = FALSE)
}
\author{
Jordan L. Prendez, \email{jordanprendez@gmail.com}
}
\keyword{IR}
\keyword{MHEC,}
\keyword{University}
\keyword{maryland,}
\keyword{of}
\keyword{system}
