% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_labels.R
\name{import_labels}
\alias{import_labels}
\title{import labels function}
\usage{
import_labels(dataset, definitions)
}
\arguments{
\item{dataset}{= path of the file containing the raw data (i.e. 1, and 0's not "YES" and "NO"). The only file format currently accepted is .csv}

\item{definitions}{= path of the file containing definitions. The format of this file is shown below in table 1. File must be a .txt}
}
\value{
Returns dataset with attached labels
}
\description{
A function to import SPSS ®  style labels and attach them to a dataset in R
}
\examples{
import_labels(dataset="C:/Users/jprendez/Desktop/ic2015.csv", definitions="C:/Users/jprendez/Desktop/definitions.txt")
}
\author{
Jordan L. Prendez, \email{jordanprendez@gmail.com}
}
\keyword{cats}

