% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USM_labels.R
\name{usm_labels}
\alias{usm_labels}
\title{usm_labels function}
\usage{
usm_labels(dataset = NULL, label_values = TRUE, label_variables = TRUE,
  labels_output = FALSE, label_matrix = NULL, manual_label_input = NULL)
}
\arguments{
\item{dataset}{= name of the object (file) that you wish to rename with USM labels.}

\item{label_values}{= TRUE (default). TRUE outputs your dataset with MHEC value labels attached.}

\item{label_variables}{= TRUE (default). TRUE outputs your dataset with MHEC variable labels attached.}

\item{labels_output}{= FALSE (default). TRUE outputs the labels used in the function.}

\item{manual_label_input}{specify labels and definitions manually in R. [examples forthcoming]}
}
\value{
Returns dataset with USM value labels and or variable names attached (depending on selected option).
}
\description{
A function to attach USM-MHEC variable labels to a file currently loaded to memory.
Specifically, this file renames database style names to those more appropriate for reading and or tables.
}
\examples{
usm_labels(dataset=my_dataset, label_values = TRUE, label_variables = FALSE)
data_def<- c("var.name_IDType"   , 1, "Student", 2, "faculty", 3, "staff","var.name_USCitizen", 1, "Yes", 2, "No", "var.name_Degree", 40, "BA", 60, "MA", 81, "AA")
usm_labels(dataset=res2, label_variables = FALSE, label_values=FALSE, manual_label_input=data_def)
}
\seealso{
\href{https://github.com/nietsnel/inst.research/blob/master/README.md}{Online example}
}
\author{
Jordan L. Prendez, \email{jordanprendez@gmail.com}
}
\keyword{IR}
\keyword{MHEC,}
\keyword{University}
\keyword{maryland,}
\keyword{of}
\keyword{system}
